Support running an animation N times (#19)
Remove `TweeningType` and split its functionalities between a new `RepeatCount`
controlling the number of repeats of an animation on one hand, and
`RepeatStrategy` controlling the way an animation restarts after a loop ended
on the other hand. This allows more granular control on the type of playback.

Remove the `tweening_type` parameter from `Tween<T>::new()` and replace it with
builder methods `with_repeat_count()` and `with_repeat_strategy()`.

Remove `is_looping()` from all tweenables, which was not implemented for most
of them anyway.